Update the required packages list
This is a comprehensive update to the packages list, both prod and dev,
primarily as this is one big experiment of an application. The Imagick
extension has been added as a requirement, as it's central to the
application.

Then, laminas inputfilter and validator were added to simplify
validating the uploaded image. mezzio-twigrenderer was added to provide
a simple view layer, mostly for forms and for rendering the available
images. And lastly monolog was added to provide some simple application
instrumentation.