Refactor the Image entity
Two changes needed to be made to the entity. Firstly, strict typing
doesn't work as the BLOB data returned from the database is done so as a
PHP resource type, which isn't currently able to be typed.

Secondly, getData() was refactored to make it easier to render the image
data inside an img tag. Sure, this likely isn't the best way, and should
be done in the view layer, but from the perspective of a tutorial, it's
fine.