UI Update
This change covers a set of UI updates that make the UI, overall, a bit
cleaner and more complete across both desktop and mobile breakpoints.

Specifically, it updates the spacing around a range of elements so that
they have a proper balance of space. It shortens the text of the upload
button in the main navigation if it's displayed on a smaller resolution
but shows the full version otherwise.