Add the ability to optimise the uploaded image
This change adds the ability for the user to choose whether they want to
optimise the image, post-upload, by converting it from whatever format
it was to AVIF format. The reason for doing this is that, while the
maximum allowed image size is 5MB, which on its own isn't that big, it
doesn't take long for the payload of the site to increase significantly.

What's more, as AVIF files can be around a tenth of the size of other
formats, and I'm loving more about images, image optimisation, and
Imagick in general, plus adding this functionality would increase the
value of the accompanying tutorial, this functionality has been added.